only attempt to delete if not already done so

+++ R/logfile.R
@@ -39,11 +39,15 @@ Logfile <- R6::R6Class(
       })
     },
     remove = function() {
-      self$active <- FALSE
-      removed <- file.remove(self$path)
-      lock_removed <- file.remove(self$lock)
-      cat("Deleted logfile at", self$path, "\n")
-      invisible(removed & lock_removed)
+      if (self$active) {
+        removed <- file.remove(self$path)
+        lock_removed <- file.remove(self$lock)
+        cat("Deleted logfile at", self$path, "\n")
+        self$active <- FALSE
+        invisible(removed & lock_removed)
+      } else {
+        cat("Logfile already deleted; nothing to do.\n")
+      }
     },
     print = function(...) {
       if (self$active) {
